올가미 상호작용전 flow까지 구현
올가미와 상호작용하기 직전  flow까지 구현
생성될 마네킹과 눈의 개수는 GameManager 에서 설정.
마네킹과 눈의 배치는 씬에서 직접 생성후 각 SpawnerSet자식에 두어 설정.
GameManager 각 설정 변수에 마우스 올렸을때 설명이 나타나도록 툴팁추가.

문,양초 오류 수

+++ Assets/03_Script/GameManager.cs
@@ -28,12 +28,12 @@ public partial class GameManager : MonoBehaviour
         //gameState = 0;
         middleGameState = false;
     }
-    IEnumerator DisableControler(int seconds)
+    IEnumerator DisableControler(float seconds)
     {
         yield return new WaitForSeconds(seconds);
         playerControler.enabled = true;
     }
-    public void DisableControlerInSec(int seconds)
+    public void DisableControlerInSec(float seconds)
     {
         playerControler.enabled = false;
         StartCoroutine("DisableControler", seconds);
